+++ shv/__init__.py
@@ -1,6 +1,18 @@
 """Python implementation of Silicon Heaven."""
 from . import chainpack, cpcontext, cpon
 from .clientconnection import ClientConnection
+from .rpcerrors import (
+    RpcError,
+    RpcErrorCode,
+    RpcInternalError,
+    RpcInvalidParamsError,
+    RpcInvalidRequestError,
+    RpcMethodCallCancelledError,
+    RpcMethodCallExceptionError,
+    RpcMethodCallTimeoutError,
+    RpcMethodNotFoundError,
+    RpcParseError,
+)
 from .rpcclient import RpcClient
 from .rpcmessage import RpcMessage
 from .rpcprotocol import RpcProtocol
@@ -11,10 +23,27 @@ __all__ = [
     "chainpack",
     "cpcontext",
     "cpon",
+    # rpcclient
     "RpcClient",
+    # rpcserver
     "RpcServer",
+    # rpcprotocol
     "RpcProtocol",
+    # rpcmessage
     "RpcMessage",
+    # rpcvalue
     "RpcValue",
+    # rpcerror
+    "RpcErrorCode",
+    "RpcError",
+    "RpcInternalError",
+    "RpcInvalidParamsError",
+    "RpcInvalidRequestError",
+    "RpcMethodCallCancelledError",
+    "RpcMethodCallExceptionError",
+    "RpcMethodCallTimeoutError",
+    "RpcMethodNotFoundError",
+    "RpcParseError",
+    # clientconnection
     "ClientConnection",
 ]