+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `RpcError` that represents SHV RPC errors
+
+### Changed
+- `RpcMessage.ErrorCode` moved and renamed to `RpcErrorCode`
+- `RpcClient.read_rpc_message` now raises `RpcError` instead of
+  `RpcClient.MethodCallError`
+
+
 ## [0.1.1] - 2023-05-22
 ### Fixed
 - Handler lookup in `ClientConnection` now correctly selects longest match